4.6.2.3d
Minor bug fixes:
- Mary can now protect herself against ricochet on the first turn
- Clio now takes into account the buff from Fairies
- Yeti now buffs attack with the correct value and on the correct turn
- Galla now kills units before they attack (so it's finally a win against w9)

+++ cosmosData.cpp
@@ -145,7 +145,7 @@ Monster::Monster(const Monster & baseHero, int aLevel, int aPromo) :
                             this->skill.amount = floor(this->skill.amount * (double) aLevel);
                             break;
         case BUFFUP_L:      this->skill.skillType = BUFFUP;
-                            this->skill.amount = 9 * floor((double) aLevel / 9);
+                            this->skill.amount = 9 * floor(aLevel / 9);
                             break;
         case FLATLEP_L:     this->skill.skillType = FLATLEP;
                             this->skill.amount = this->skill.amount * floor(aLevel / 9);
